Move `MapAscl` to `map_loading` crate.
