+++ crate/loading/Cargo.toml
@@ -42,6 +42,7 @@ kinematic_model = { path = "../kinematic_model" }
 loading_model = { path = "../loading_model" }
 loading_spi = { path = "../loading_spi" }
 log = "0.4.8"
+map_loading = { path = "../map_loading" }
 map_model = { path = "../map_model" }
 map_selection_ui_model = { path = "../map_selection_ui_model" }
 mirrored_model = { path = "../mirrored_model" }
