+++ crate/loading/src/system/asset_sequence_component_loading_system.rs
@@ -7,6 +7,7 @@ use kinematic_model::loaded::ObjectAccelerationSequenceHandles;
 use loading_model::loaded::LoadStage;
 use loading_spi::{AssetLoadingResources, SequenceComponentLoadingResources};
 use log::debug;
+use map_loading::MapAscl;
 use object_loading::ObjectAscl;
 use sequence_model::loaded::WaitSequenceHandles;
 use spawn_model::loaded::SpawnsSequenceHandles;
@@ -16,10 +17,6 @@ use sprite_model::loaded::{
 use ui_loading::UiAscl;
 
 use crate::{AssetPartLoader, AssetPartLoadingSystem};
-
-pub use self::map_ascl::MapAscl;
-
-mod map_ascl;
 
 /// Loads asset sequence components.
 pub type AssetSequenceComponentLoadingSystem = AssetPartLoadingSystem<AssetSequenceComponentLoader>;
