+++ crate/map_loading/src/lib.rs
@@ -2,6 +2,7 @@
 
 //! Processes map configuration into the loaded map model.
 
-pub use crate::map_loading_bundle::MapLoadingBundle;
+pub use crate::{map_ascl::MapAscl, map_loading_bundle::MapLoadingBundle};
 
+mod map_ascl;
 mod map_loading_bundle;
